refactor(findAllByUserId): getOpenTime, getCloseTime, isActive 변수들 내부 캡슐화

+++ backend/src/main/java/com/woowacourse/moragora/domain/participant/ParticipantAndCount.java
@@ -4,5 +4,5 @@ public interface ParticipantAndCount {
 
     Participant getParticipant();
 
-    Long getTardyCount();
+    Integer getTardyCount();
 }
